+++ src/main/java/lotto/domain/LottoShop.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.LottoPrice;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
