+++ src/main/java/lotto/domain/exception/NotValidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.exception;
 
 public class NotValidLottoNumberException extends RuntimeException{
